moved Translation unit tests resources to separate dir
