+++ packages/framework/tests/Unit/Component/Translation/Resources/ConstraintClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\FrameworkBundle\Unit\Component\Translation;
+namespace Tests\FrameworkBundle\Unit\Component\Translation\Resources;
 
 use Symfony\Component\Validator\Constraint;
