+++ packages/framework/tests/Unit/Component/Translation/Resources/NonConstraintClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\FrameworkBundle\Unit\Component\Translation;
+namespace Tests\FrameworkBundle\Unit\Component\Translation\Resources;
 
 class NonConstraintClass
 {
