+++ packages/framework/tests/Unit/Component/Translation/Resources/test.js
